requirements.txt
Pillow
numpy
opencv-python-headless
scipy
matplotlib
folium
PyPDF2
pytest
